requirements.txt
# Human Design Calculator Dependencies
streamlit>=1.28.0
pyswisseph>=2.10.3.2
plotly>=5.18.0
pandas>=2.0.0
pytz>=2023.3
geopy>=2.4.1
timezonefinder>=6.2.0
